# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClinicalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospitals:
    get:
      operationId: listHospitals
      tags: [hospitals]
      summary: List hospitals
      responses:
        "200":
          description: List of hospitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hospital"
    post:
      operationId: createHospital
      tags: [hospitals]
      summary: Create hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /hospitals/{id}:
    patch:
      operationId: updateHospital
      tags: [hospitals]
      summary: Update hospital
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /hospitals/{id}/departments:
    get:
      operationId: listDepartments
      tags: [hospitals]
      summary: List departments for a hospital
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [hospitals]
      summary: Create department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  # ── CLINICAL CASES ────────────────────────────────────────────────────────
  /cases:
    get:
      operationId: listClinicalCases
      tags: [cases]
      summary: List required clinical cases
      responses:
        "200":
          description: Clinical cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClinicalCase"
    post:
      operationId: createClinicalCase
      tags: [cases]
      summary: Create a new case type (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicalCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /cases/{id}:
    patch:
      operationId: updateClinicalCase
      tags: [cases]
      summary: Update clinical case type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicalCaseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalCase"
  /case-completions:
    get:
      operationId: listCaseCompletions
      tags: [cases]
      summary: List case completion submissions
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected]
        - name: scheduleId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Case completions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseCompletion"
    post:
      operationId: submitCaseCompletion
      tags: [cases]
      summary: Student submits a case completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCompletionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCompletion"
  /case-completions/{id}/verify:
    patch:
      operationId: verifyCaseCompletion
      tags: [cases]
      summary: Scheduler verifies a case submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCompletion"
  /case-completions/{id}/reject:
    patch:
      operationId: rejectCaseCompletion
      tags: [cases]
      summary: Scheduler rejects a case submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseRejection"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCompletion"

  # ── SCHEDULES ─────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List my notifications
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role, firstName, lastName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        ciProfile:
          $ref: "#/components/schemas/CIProfile"
...
    StudentProfile:
      type: object
      required: [id, userId, studentNumber]
      properties:
        id:
          type: string
        userId:
          type: string
        studentNumber:
          type: string
        yearLevel:
          type: ["integer", "null"]
        section:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
        totalHoursRequired:
          type: integer
        createdAt:
          type: string
          format: date-time
    CIProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
        userId:
          type: string
        employeeId:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]

    # ── STUDENT DETAILS ────────────────────────────────────────────────────
...
    Hospital:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        attendanceRadius:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
    HospitalInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        attendanceRadius:
          type: integer
        isActive:
          type: boolean
    Department:
      type: object
      required: [id, hospitalId, name, isActive]
      properties:
        id:
          type: string
        hospitalId:
          type: string
        name:
          type: string
        code:
          type: ["string", "null"]
        isActive:
          type: boolean
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string
        isActive:
          type: boolean

    # ── CLINICAL CASES ─────────────────────────────────────────────────────
    ClinicalCase:
      type: object
      required: [id, name, category, requiredCount, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        requiredCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ClinicalCaseInput:
      type: object
      required: [name, category, requiredCount]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        requiredCount:
          type: integer
        isActive:
          type: boolean
    CaseCompletion:
      type: object
      required: [id, studentId, clinicalCaseId, status, submittedAt]
      properties:
        id:
          type: string
        studentId:
          type: string
        clinicalCaseId:
          type: string
        scheduleId:
          type: ["string", "null"]
        hospitalId:
          type: ["string", "null"]
        departmentId:
          type: ["string", "null"]
        submittedAt:
          type: string
          format: date-time
        verifiedAt:
          type: ["string", "null"]
        verifiedBy:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
        rejectionReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        clinicalCase:
          $ref: "#/components/schemas/ClinicalCase"
        student:
          $ref: "#/components/schemas/User"
        hospital:
          $ref: "#/components/schemas/Hospital"
        department:
          $ref: "#/components/schemas/Department"
    CaseCompletionInput:
      type: object
      required: [clinicalCaseId, scheduleId, hospitalId, departmentId]
      properties:
        clinicalCaseId:
          type: string
        scheduleId:
          type: string
        hospitalId:
          type: string
        departmentId:
          type: string
        notes:
          type: string
        photoUrl:
          type: string
    CaseRejection:
      type: object
      required: [rejectionReason]
      properties:
        rejectionReason:
          type: string

    # ── SCHEDULES ──────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        relatedEntity:
          type: ["string", "null"]
        relatedId:
          type: ["string", "null"]
        isRead:
          type: boolean
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── ANNOUNCEMENTS ──────────────────────────────────────────────────────
...